Move some internal functions in crypt.go into an internal package
